requirements.txt
# Core dependencies
streamlit>=1.22
PyYAML>=6.0
python-dotenv>=1.0.0

# LLM and embeddings
openai>=1.3.0
tiktoken>=0.5.1

# YouTube transcript + metadata
youtube-transcript-api>=0.4.5
pytube>=15.0.0
yt-dlp>=2024.4.9

# Embedding + search
faiss-cpu>=1.7.4
numpy>=1.23.0

